Tags NOT yaml. In DB.  Yelp API calls work.  DB seeded with venues.  Weird tags

+++ db/migrate/20130803170048_create_venues.rb
@@ -7,10 +7,12 @@ class CreateVenues < ActiveRecord::Migration
       t.integer :zip
       t.string :city
       t.string :state
-      t.integer :rating
+      t.float :rating
       t.string :img_url
       t.string :biz_url
       t.timestamps
+
+
     end
   end
 end
